RavenDB-21771 - create an abstraction for the Event Listener
