+++ src/Raven.Server/EventListener/EventType.cs
@@ -1,0 +1,10 @@
+﻿namespace Raven.Server.EventListener;
+
+public enum EventType
+{
+    GC,
+    GCSuspend,
+    GCRestart,
+    GCFinalizers,
+    Contention
+}
